Move ui notification to service layer
