+++ routers/init.go
@@ -50,6 +50,7 @@ import (
 	repo_service "code.gitea.io/gitea/services/repository"
 	"code.gitea.io/gitea/services/repository/archiver"
 	"code.gitea.io/gitea/services/task"
+	"code.gitea.io/gitea/services/uinotification"
 	"code.gitea.io/gitea/services/webhook"
 )
 
@@ -121,6 +122,7 @@ func InitWebInstalled(ctx context.Context) {
 	mailer.NewContext(ctx)
 	mustInit(cache.NewContext)
 	notification.NewContext()
+	mustInit(uinotification.Init)
 	mustInit(archiver.Init)
 
 	highlight.NewContext()
